Moved DAT classes under own namespace.
