+++ src/NTwain/NTwain.csproj
@@ -37,4 +37,9 @@
     </EmbeddedResource>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Triplets\Image\" />
+    <Folder Include="Triplets\Audio\" />
+  </ItemGroup>
+
 </Project>
