+++ src/NTwain/Triplets/Control/Callback2.cs
@@ -1,7 +1,7 @@
 ﻿using NTwain.Data;
 using NTwain.Internals;
 
-namespace NTwain.Triplets
+namespace NTwain.Triplets.Control
 {
     sealed class Callback2 : BaseTriplet
     {
