+++ src/NTwain/Triplets/Control/Parent.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace NTwain.Triplets
+namespace NTwain.Triplets.Control
 {
     sealed partial class Parent : BaseTriplet
     {
