+++ src/NTwain/Triplets/DGControl.cs
@@ -1,8 +1,10 @@
 ﻿using NTwain.Data;
+using NTwain.Triplets.Control;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using DeviceEvent = NTwain.Triplets.Control.DeviceEvent;
 
 namespace NTwain.Triplets
 {
